+++ tmtccmd/pus/__init__.py
@@ -1,7 +1,7 @@
 from enum import IntEnum
 from typing import Optional
 
-from .pus_11_tc_sched import Subservices as Pus11Subservices
+from .pus_11_tc_sched import Subservice as Pus11Subservices
 from spacepackets.ecss import PusTelecommand
 from spacepackets.ecss.pus_1_verification import RequestId
 import spacepackets.ecss.pus_1_verification as pus_1