+++ tmtccmd/pus/pus_201_fsfw_health.py
@@ -1,7 +1,7 @@
 import enum
 
 
-class Subservices(enum.IntEnum):
+class Subservice(enum.IntEnum):
     TC_SET_HEALTH = 1
     TM_HEALTH_SET = 2
     TC_ANNOUNCE_HEALTH = 3